Fix tyipe

+++ Stonks/Stonks/Services/Data/Protocols/DataProtocols.swift
@@ -37,7 +37,7 @@ protocol StockDataServiceInput {
 protocol StockHistoryDataServiceInput {
     func createHistoryStock(name: String,
                             symbol: String,
-                            price: Double,
+                            price: Decimal,
                             type: TypeOfAction)
     func getAllStocks() -> [StockHistory]?
     func getStocks(with type: TypeOfAction?, sortby: SortBy?) -> [StockHistory]?
